backend/requirements.txt
amqp==5.3.1
asgiref==3.10.0        
async-timeout==5.0.1   
billiard==4.2.2        
celery==5.5.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==5.2.7
django-cors-headers==4.9.0
django-redis==6.0.0
django_celery_results==2.6.0
djangorestframework==3.16.1
et_xmlfile==2.0.0
kombu==5.5.4
numpy==2.3.4
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
redis==6.4.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.14